kubeadm: Move app/master into a separate phase directory
