+++ cmd/kubeadm/app/cmd/init.go
@@ -32,10 +32,10 @@ import (
 	kubeadmapiext "k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/v1alpha1"
 	"k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/validation"
 	kubeadmconstants "k8s.io/kubernetes/cmd/kubeadm/app/constants"
-	kubemaster "k8s.io/kubernetes/cmd/kubeadm/app/master"
 	addonsphase "k8s.io/kubernetes/cmd/kubeadm/app/phases/addons"
 	apiconfigphase "k8s.io/kubernetes/cmd/kubeadm/app/phases/apiconfig"
 	certphase "k8s.io/kubernetes/cmd/kubeadm/app/phases/certs"
+	controlplanephase "k8s.io/kubernetes/cmd/kubeadm/app/phases/controlplane"
 	kubeconfigphase "k8s.io/kubernetes/cmd/kubeadm/app/phases/kubeconfig"
 	tokenphase "k8s.io/kubernetes/cmd/kubeadm/app/phases/token"
 	"k8s.io/kubernetes/cmd/kubeadm/app/preflight"
@@ -211,28 +211,18 @@ func (i *Init) Run(out io.Writer) error {
 	}
 
 	// PHASE 3: Bootstrap the control plane
-	if err := kubemaster.WriteStaticPodManifests(i.cfg); err != nil {
+	if err := controlplanephase.WriteStaticPodManifests(i.cfg); err != nil {
 		return err
 	}
 
 	adminKubeConfigPath := filepath.Join(kubeadmapi.GlobalEnvParams.KubernetesDir, kubeadmconstants.AdminKubeConfigFileName)
-	client, err := kubemaster.CreateClientAndWaitForAPI(adminKubeConfigPath)
+	client, err := kubeadmutil.CreateClientAndWaitForAPI(adminKubeConfigPath)
 	if err != nil {
 		return err
 	}
 
 	if err := apiconfigphase.UpdateMasterRoleLabelsAndTaints(client); err != nil {
 		return err
-	}
-
-	// Is deployment type self-hosted?
-	if i.cfg.SelfHosted {
-		// Temporary control plane is up, now we create our self hosted control
-		// plane components and remove the static manifests:
-		fmt.Println("[self-hosted] Creating self-hosted control plane...")
-		if err := kubemaster.CreateSelfHostedControlPlane(i.cfg, client); err != nil {
-			return err
-		}
 	}
 
 	// PHASE 4: Set up the bootstrap tokens
@@ -269,6 +259,16 @@ func (i *Init) Run(out io.Writer) error {
 
 	if err := addonsphase.CreateEssentialAddons(i.cfg, client); err != nil {
 		return err
+	}
+
+	// Is deployment type self-hosted?
+	if i.cfg.SelfHosted {
+		// Temporary control plane is up, now we create our self hosted control
+		// plane components and remove the static manifests:
+		fmt.Println("[self-hosted] Creating self-hosted control plane...")
+		if err := controlplanephase.CreateSelfHostedControlPlane(i.cfg, client); err != nil {
+			return err
+		}
 	}
 
 	ctx := map[string]string{
