+++ cmd/kubeadm/app/util/BUILD
@@ -11,14 +11,19 @@ load(
 go_library(
     name = "go_default_library",
     srcs = [
+        "apiclient.go",
         "error.go",
         "template.go",
         "version.go",
     ],
     tags = ["automanaged"],
     deps = [
+        "//cmd/kubeadm/app/constants:go_default_library",
         "//cmd/kubeadm/app/preflight:go_default_library",
+        "//cmd/kubeadm/app/util/kubeconfig:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/errors:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/util/wait:go_default_library",
+        "//vendor/k8s.io/client-go/kubernetes:go_default_library",
     ],
 )
